arch(testing): apply Phase 1B r2 delta — driver/ sub-folder, not merged file
Per user OK-to-advance delta on #541: replace `_shared/wire-driver.ts`
(merged file) with `_shared/driver/` sub-folder containing
`test-client.ts`, `test-server.ts`, and `index.ts` barrel as siblings.

Rationale (user): driver/ is a coherent sub-package, not a forced
merger. test-client + test-server are wire-driver harness pieces that
share a journal type and an error channel; they belong together in
driver/ but stay as separate files so each role reads as one file.

Also fixes Data.Case stub error (effect 3.21 has no Data.Case export);
both stubs now use the production-grade Data.TaggedError pattern they
target on port.

Refs #541. Plan delta in #541 comment.
